Can now handle widespread cells missing from simulation.
Does this by setting corresponding crop fractions to zero, warning with affected area. If a country's area excluded in this way for a certain crop is > 10% (settable as ctry_excluded_area_thresh), then that country will be excluded from the calibration. For countries that experience some reduction but are not excluded, note that FAO data are not reduced in any way! Because the calibration happens on yield (not total production), this effectively assumes that all excluded gridcells in a country are perfectly average.