e2p: Reworked handling of cells where baseline is zero but future is positive.
Previously---and this is preserved in now-unused functions e2p_get_deltas() and e2p_apply_deltas()---infinite deltas were simply interpolated over. Infinite deltas happen when the EMULATOR baseline (NOT the USED baseline, which is the one the deltas will be applied to) is zero. Now, in cells where this is the case, all deltas beginning the first nonzero future emulated year are calculated relative to the first nonzero future emulated year. At the same time, in such cells, the USED baseline is INTERPOLATED IN FOR SOME REASON?